+++ engine/src/main/java/org/camunda/bpm/engine/impl/HistoricJobLogQueryImpl.java
@@ -47,8 +47,8 @@ public class HistoricJobLogQueryImpl extends AbstractQuery<HistoricJobLogQuery, 
   protected String processDefinitionKey;
   protected String deploymentId;
   protected JobState state;
-  protected Integer jobPriorityHigherThanOrEqual;
-  protected Integer jobPriorityLowerThanOrEqual;
+  protected Long jobPriorityHigherThanOrEqual;
+  protected Long jobPriorityLowerThanOrEqual;
 
   public HistoricJobLogQueryImpl() {
   }
@@ -139,12 +139,12 @@ public class HistoricJobLogQueryImpl extends AbstractQuery<HistoricJobLogQuery, 
     return this;
   }
 
-  public HistoricJobLogQuery jobPriorityHigherThanOrEquals(int priority) {
+  public HistoricJobLogQuery jobPriorityHigherThanOrEquals(long priority) {
     this.jobPriorityHigherThanOrEqual = priority;
     return this;
   }
 
-  public HistoricJobLogQuery jobPriorityLowerThanOrEquals(int priority) {
+  public HistoricJobLogQuery jobPriorityLowerThanOrEquals(long priority) {
     this.jobPriorityLowerThanOrEqual = priority;
     return this;
   }
