+++ engine/src/main/java/org/camunda/bpm/engine/impl/persistence/entity/JobManager.java
@@ -288,7 +288,7 @@ public class JobManager extends AbstractManager {
     getDbEntityManager().update(JobEntity.class, "updateFailedJobRetriesByParameters", parameters);
   }
 
-  public void updateJobPriorityByDefinitionId(String jobDefinitionId, int priority) {
+  public void updateJobPriorityByDefinitionId(String jobDefinitionId, long priority) {
     Map<String, Object> parameters = new HashMap<String, Object>();
     parameters.put("jobDefinitionId", jobDefinitionId);
     parameters.put("priority", priority);
